Add files via upload
This project begins by introducing some simple real-life examples for
regression. From a brief introduction to most of the concepts used in regression to
hands-on experience, this project will give you enough understanding to apply those in
real-world problems. With the help of the background developed, you will code your
regression model in python.